move TRIAL_STATUS_TYPES, SCAN_MODES, EMPLOYER_OPTIONS, PRACTITIONER_TYPE_OPTIONS, ADMISSIONS_STATUS_OPTIONS
