+++ shared/src/business/entities/EntityConstants.js
@@ -666,15 +666,53 @@ const CHAMBERS_SECTIONS = sortBy([
   WELLS_CHAMBERS_SECTION,
 ]);
 
+const TRIAL_STATUS_TYPES = [
+  'Set for Trial',
+  'Dismissed',
+  'Continued',
+  'Rule 122',
+  'A Basis Reached',
+  'Settled',
+  'Recall',
+  'Taken Under Advisement',
+];
+
+const SCAN_MODES = {
+  DUPLEX: 'duplex',
+  FEEDER: 'feeder',
+  FLATBED: 'flatbed',
+};
+
+const EMPLOYER_OPTIONS = ['IRS', 'DOJ', 'Private'];
+
+const PRACTITIONER_TYPE_OPTIONS = ['Attorney', 'Non-Attorney'];
+
+const ADMISSIONS_STATUS_OPTIONS = [
+  'Active',
+  'Suspended',
+  'Disbarred',
+  'Resigned',
+  'Deceased',
+  'Inactive',
+];
+
+const DEFAULT_PROCEDURE_TYPE = PROCEDURE_TYPES[0];
+
+const CASE_SEARCH_MIN_YEAR = 1986;
+const CASE_SEARCH_PAGE_SIZE = 5;
+
 module.exports = {
   ADC_SECTION,
   ADMISSIONS_SECTION,
+  ADMISSIONS_STATUS_OPTIONS,
   ANSWER_CUTOFF_AMOUNT_IN_DAYS,
   ANSWER_CUTOFF_UNIT,
   ANSWER_DOCUMENT_CODES,
   AUTOMATIC_BLOCKED_REASONS,
   BUSINESS_TYPES,
   CASE_CAPTION_POSTFIX,
+  CASE_SEARCH_MIN_YEAR,
+  CASE_SEARCH_PAGE_SIZE,
   CASE_STATUS_TYPES,
   CASE_TYPES,
   CASE_TYPES_MAP,
@@ -685,6 +723,7 @@ module.exports = {
   CONTACT_CHANGE_DOCUMENT_TYPES,
   COUNTRY_TYPES,
   COURT_ISSUED_EVENT_CODES,
+  DEFAULT_PROCEDURE_TYPE,
   DOCKET_NUMBER_MATCHER,
   DOCKET_NUMBER_SUFFIXES,
   DOCKET_SECTION,
@@ -694,6 +733,7 @@ module.exports = {
   DOCUMENT_INTERNAL_CATEGORY_MAP,
   DOCUMENT_NOTICE_EVENT_CODES,
   DOCUMENT_RELATIONSHIPS,
+  EMPLOYER_OPTIONS,
   ESTATE_TYPES,
   EVENT_CODES,
   FILING_TYPES,
@@ -712,8 +752,10 @@ module.exports = {
   PAYMENT_STATUS,
   PETITIONS_SECTION,
   PRACTITIONER_ASSOCIATION_DOCUMENT_TYPES,
+  PRACTITIONER_TYPE_OPTIONS,
   PROCEDURE_TYPES,
   ROLES,
+  SCAN_MODES,
   SCENARIOS,
   SECTIONS,
   SERVICE_INDICATOR_TYPES,
@@ -732,5 +774,6 @@ module.exports = {
   TRIAL_CITY_STRINGS,
   TRIAL_CLERKS_SECTION,
   TRIAL_LOCATION_MATCHER,
+  TRIAL_STATUS_TYPES,
   US_STATES,
 };
