+++ shared/src/business/entities/Practitioner.js
@@ -1,5 +1,11 @@
 const joi = require('@hapi/joi');
 const {
+  ADMISSIONS_STATUS_OPTIONS,
+  EMPLOYER_OPTIONS,
+  PRACTITIONER_TYPE_OPTIONS,
+  ROLES,
+} = require('./EntityConstants');
+const {
   joiValidationDecorator,
 } = require('../../utilities/JoiValidationDecorator');
 const {
@@ -8,19 +14,7 @@ const {
   VALIDATION_ERROR_MESSAGES: USER_VALIDATION_ERROR_MESSAGES,
 } = require('./User');
 const { getTimestampSchema } = require('../../utilities/dateSchema');
-const { ROLES } = require('./EntityConstants');
 const joiStrictTimestamp = getTimestampSchema();
-
-const EMPLOYER_OPTIONS = ['IRS', 'DOJ', 'Private'];
-const PRACTITIONER_TYPE_OPTIONS = ['Attorney', 'Non-Attorney'];
-const ADMISSIONS_STATUS_OPTIONS = [
-  'Active',
-  'Suspended',
-  'Disbarred',
-  'Resigned',
-  'Deceased',
-  'Inactive',
-];
 
 /**
  * constructor
@@ -190,12 +184,8 @@ joiValidationDecorator(
 );
 
 Practitioner.validationName = 'Practitioner';
-
-Practitioner.PRACTITIONER_TYPE_OPTIONS = PRACTITIONER_TYPE_OPTIONS;
-Practitioner.EMPLOYER_OPTIONS = EMPLOYER_OPTIONS;
 Practitioner.validationRules = practitionerValidation;
 Practitioner.VALIDATION_ERROR_MESSAGES = VALIDATION_ERROR_MESSAGES;
-Practitioner.ADMISSIONS_STATUS_OPTIONS = ADMISSIONS_STATUS_OPTIONS;
 
 /**
  * returns the full concatenated name for the given practitioner data
