+++ shared/src/business/entities/Scan.js
@@ -5,6 +5,7 @@ const {
 const { createISODateString } = require('../utilities/DateHandler');
 const { getTimestampSchema } = require('../../utilities/dateSchema');
 const { remove } = require('lodash');
+
 const joiStrictTimestamp = getTimestampSchema();
 /**
  * constructor
@@ -21,11 +22,6 @@ function Scan({ applicationContext, rawScan }) {
 }
 
 Scan.validationName = 'Scan';
-Scan.SCAN_MODES = {
-  DUPLEX: 'duplex',
-  FEEDER: 'feeder',
-  FLATBED: 'flatbed',
-};
 
 /**
  * adds a batch to the current scan
