+++ shared/src/business/entities/cases/CaseSearch.js
@@ -2,9 +2,7 @@ const joi = require('@hapi/joi');
 const {
   joiValidationDecorator,
 } = require('../../../utilities/JoiValidationDecorator');
-
-CaseSearch.CASE_SEARCH_MIN_YEAR = 1986;
-CaseSearch.CASE_SEARCH_PAGE_SIZE = 5;
+const { CASE_SEARCH_MIN_YEAR } = require('../EntityConstants');
 
 CaseSearch.validationName = 'CaseSearch';
 
@@ -16,7 +14,7 @@ CaseSearch.validationName = 'CaseSearch';
  */
 function CaseSearch(rawProps) {
   this.petitionerName = rawProps.petitionerName;
-  this.yearFiledMin = rawProps.yearFiledMin || CaseSearch.CASE_SEARCH_MIN_YEAR;
+  this.yearFiledMin = rawProps.yearFiledMin || CASE_SEARCH_MIN_YEAR;
   this.yearFiledMax = rawProps.yearFiledMax || undefined;
   this.petitionerState = rawProps.petitionerState || undefined;
   this.countryType = rawProps.countryType || undefined;
