+++ web-api/src/applicationContext.js
@@ -1055,7 +1055,7 @@ module.exports = (appContextUser = {}) => {
         status => status !== CASE_STATUS_TYPES.closed,
       ),
       ORDER_TYPES_MAP: ORDER_TYPES,
-      SESSION_STATUS_GROUPS: SESSION_STATUS_GROUPS,
+      SESSION_STATUS_GROUPS,
     }),
     getCurrentUser,
     getDispatchers: () => ({
