+++ web-client/src/presenter/actions/rescanBatchAction.test.js
@@ -1,12 +1,10 @@
-import { Scan } from '../../../../shared/src/business/entities/Scan';
+import { SCAN_MODES } from '../../../../shared/src/business/entities/EntityConstants';
 import { applicationContextForClient as applicationContext } from '../../../../shared/src/business/test/createTestApplicationContext';
 import { presenter } from '../presenter-mock';
 import { rescanBatchAction } from './rescanBatchAction';
 import { runAction } from 'cerebral/test';
 
 describe('rescanBatchAction', () => {
-  const { SCAN_MODES } = Scan;
-
   const successStub = jest.fn();
   const errorStub = jest.fn();
