+++ web-client/src/presenter/computeds/trialSessionWorkingCopyHelper.test.js
@@ -1,7 +1,9 @@
-import { CASE_STATUS_TYPES } from '../../../../shared/src/business/entities/EntityConstants';
-import { Case } from '../../../../shared/src/business/entities/cases/Case';
+import {
+  CASE_STATUS_TYPES,
+  STATUS_TYPES,
+  TRIAL_STATUS_TYPES,
+} from '../../../../shared/src/business/entities/EntityConstants';
 import { MOCK_CASE } from '../../../../shared/src/test/mockCase';
-import { TrialSessionWorkingCopy } from '../../../../shared/src/business/entities/trialSessions/TrialSessionWorkingCopy';
 import { applicationContext } from '../../applicationContext';
 import { runCompute } from 'cerebral/test';
 import { trialSessionWorkingCopyHelper as trialSessionWorkingCopyHelperComputed } from './trialSessionWorkingCopyHelper';
@@ -400,8 +402,8 @@ describe('trial session working copy computed', () => {
     let result = runCompute(trialSessionWorkingCopyHelper, {
       state: {
         constants: {
-          STATUS_TYPES: Case.STATUS_TYPES,
-          TRIAL_STATUS_TYPES: TrialSessionWorkingCopy.TRIAL_STATUS_TYPES,
+          STATUS_TYPES,
+          TRIAL_STATUS_TYPES,
         },
         trialSession: {
           ...TRIAL_SESSION,
